+++ src/main/java/com/mega/s1/board/notice/NoticeVO.java
@@ -9,7 +9,7 @@ public class NoticeVO extends BoardVO {
 	
 	private String separate;
 
-	private long theaterNum;
+	private int theaterNum;
 	
 	
 }
